Refatoração das funções de pacientes
