+++ routes/__init__.py
@@ -1,0 +1,3 @@
+from .pacientes.pacientes_app import pacientes_routes
+
+__all__ = ['pacientes_routes']
