+++ routes/pacientes/__init__.py
@@ -1,0 +1,5 @@
+from .create_paciente import create_paciente
+from .update_paciente import atualizar_paciente
+from .delete_paciente import excluir_paciente
+from .get_paciente import buscar_paciente
+from .get_all_pacientes import get_all_pacientes
