+++ routes/pacientes/pacientes_app.py
@@ -1,0 +1,14 @@
+from flask import Blueprint
+from .create_paciente import create_paciente
+from .update_paciente import atualizar_paciente
+from .delete_paciente import excluir_paciente
+from .get_paciente import buscar_paciente
+from .get_all_pacientes import get_all_pacientes
+
+pacientes_routes = Blueprint('pacientes_routes', __name__)
+
+pacientes_routes.route('/api/criar_paciente', methods=['POST'])(create_paciente)
+pacientes_routes.route('/api/atualizar_paciente/<cpf>', methods=['PUT'])(atualizar_paciente)
+pacientes_routes.route('/api/excluir_paciente/<cpf>', methods=['DELETE'])(excluir_paciente)
+pacientes_routes.route('/api/buscar_paciente', methods=['GET'])(buscar_paciente)
+pacientes_routes.route('/api/exibe_pacientes', methods=['GET'])(get_all_pacientes)
